run: make namespace optional
When no namespace is provided, we don't expect an answer from the
Chromecast. When an application doesn't allocate a channel, it doesn't
send a directed answer with the supported namespaces. We still get some
general broadcast, but not one directed to our requestId.

Also rework other functions to not wait for anything if we don't have a
callback.