channel: rework low-level protocol implementation
We don't use streams any more. This was quite complicated to tear off
the pipeline on errors because we had several places were data could be
stuck and `end` events not propagated (the TLS stream and the heartbeat
transformer).

Instead, use a more straightforward approach.

From an API point of view, the user has now to wait for the connect
event before sending anything.