channel: always use a requestId
The use of a requestId is not automatic. The previous commit only used a
requestId if we expect an answer. However, on some cases, like running
an application, a requestId is needed and we don't expect an
answer. Some protocols request a requestId, some other not. Let's use it
always.